requirements.txt
slack-bolt==1.18.0
slack-sdk==3.26.2
python-dotenv==1.0.1
openai==1.30.1
python-dateutil==2.9.0.post0

